refactor: renamed sbomtestfunctions to helper
